+++ tcex/profile/migrate.py
@@ -3,6 +3,7 @@
 import json
 import re
 import sys
+from typing import Optional
 
 # third-party
 import colorama as c
@@ -14,12 +15,12 @@ c.init(autoreset=True, strip=False)
 class Migrate:
     """Class for profile Migration methods management."""
 
-    def __init__(self, profile):
+    def __init__(self, profile: object):
         """Initialize Class properties."""
         self.profile = profile
 
     @property
-    def _reserved_args(self):
+    def _reserved_args(self) -> list:
         """Return a list of *all* ThreatConnect reserved arg values."""
         return [
             'api_access_id',
@@ -57,28 +58,29 @@ class Migrate:
             'tc_user_id',
         ]
 
-    def add_staging_data(self, profile_data, name, type_, value):
+    def add_staging_data(self, profile_data: dict, name: str, type_: str, value: str) -> str:
         """Create staging data and return variable value.
 
         Args:
-            profile_data (dict): The profile to update with new staging data.
-            name (str): The name of the input.
-            type_ (str): The type of input (Binary, StringArray, etc.)
-            value (str): The value to write in the staging data.
+            profile_data: The profile to update with new staging data.
+            name: The name of the input.
+            type_: The type of input (Binary, StringArray, etc.)
+            value: The value to write in the staging data.
 
         Returns:
-            [type]: [description]
+            str: The variable for the staged data.
         """
+        profile_data.setdefault('stage', {}).setdefault('kvstore', {})
         variable = self.profile.ij.create_variable(name, type_)
         profile_data['stage']['kvstore'][variable] = value
         return variable
 
     @staticmethod
-    def comment_to_comments(profile_data):
+    def comment_to_comments(profile_data: dict) -> None:
         """Migrate comment key to comments key.
 
         Args:
-            profile_data (dict): The profile data dict.
+            profile_data: The profile data dict.
         """
         # APP-80 - migrate comment/comments field
         comment = profile_data.pop('comment', None)
@@ -141,7 +143,7 @@ class Migrate:
         self.profile.write(profile_data)
 
     @staticmethod
-    def deprecated_fields(profile_data):
+    def deprecated_fields(profile_data: dict) -> None:
         """Remove deprecated fields."""
         deprecated_fields = ['version']
         for d in deprecated_fields:
@@ -158,8 +160,12 @@ class Migrate:
             except KeyError:
                 pass
 
-    def move_default_inputs(self, profile_data):
-        """Move any default args from optional or required inputs to defaults section."""
+    def move_default_inputs(self, profile_data: dict) -> None:
+        """Move any default args from optional or required inputs to defaults section.
+
+        Args:
+            profile_data: The profile data dict.
+        """
         updated_params = []  # collect all PB configs for service Apps
         for profile_inputs in self.profile.profile_inputs:
             # ADI-1376 - handle tcex default args
@@ -182,22 +188,22 @@ class Migrate:
             profile_data['inputs'] = updated_params[0]
 
     @staticmethod
-    def permutation_output_variables(profile_data):
+    def permutation_output_variables(profile_data: dict) -> None:
         """Remove permutation_output_variables field.
 
         Args:
-            profile_data (dict): The profile data dict.
+            profile_data: The profile data dict.
         """
         try:
             del profile_data['permutation_output_variables']
         except KeyError:
             pass
 
-    def service_config_inputs(self, profile_data):
+    def service_config_inputs(self, profile_data: dict) -> None:
         """Change flat config inputs to include required/options.
 
         Args:
-            profile_data (dict): The profile data dict.
+            profile_data: The profile data dict.
         """
         for configs in profile_data.get('configs', []):
             config = configs.get('config', {})
@@ -212,8 +218,8 @@ class Migrate:
             # iterate over defined inputs
             for k, v in config.items():
                 input_data = self.profile.ij.params_dict.get(k)
+                input_type = 'optional'
                 if input_data is not None:
-                    input_type = 'optional'
                     if input_data.get('required') is True:
                         input_type = 'required'
 
@@ -224,15 +230,23 @@ class Migrate:
             configs['config'] = config_inputs
 
     @staticmethod
-    def create_staged_section(profile_data):
-        """Create the Stage and kvstore section in the profile if not currently present"""
+    def create_staged_section(profile_data: dict) -> None:
+        """Create the Stage and kvstore section in the profile if not currently present.
+
+        Args:
+            profile_data: The profile data dict.
+        """
         if 'stage' not in profile_data.keys():
             profile_data['stage'] = {'kvstore': {}}
         if 'kvstore' not in profile_data.get('stage').keys():
             profile_data['stage']['kvstore'] = []
 
-    def stage_inputs(self, profile_data):
-        """Stage any non-staged profile data."""
+    def stage_inputs(self, profile_data: dict) -> None:
+        """Stage any non-staged profile data.
+
+        Args:
+            profile_data: The profile data dict.
+        """
         if self.profile.ij.runtime_level.lower() != 'playbook':
             # staging is only required for PB Apps
             return
@@ -276,45 +290,43 @@ class Migrate:
                 profile_data['inputs'][input_type][k] = v
 
     @staticmethod
-    def stage_redis_name(profile_data):
+    def stage_redis_name(profile_data: dict) -> None:
         """Update stage.redis to stage.kvstore
 
         This change updates the previous value of redis with a
         more generic value of kvstore for staged data.
 
         Args:
-            profile_data (dict): The current profile data dict.
+            profile_data: The current profile data dict.
 
         Returns:
             dict: The update profile dict.
         """
         if profile_data.get('stage') is None:
-            return profile_data
+            return
 
-        kvstore_data = profile_data['stage'].get('redis', None)
+        kvstore_data: Optional[dict] = profile_data['stage'].get('redis', None)
         if kvstore_data is not None:
             del profile_data['stage']['redis']
             profile_data['stage']['kvstore'] = kvstore_data
 
-        return None
-
     @staticmethod
-    def stage_threatconnect_data(profile_data):
+    def stage_threatconnect_data(profile_data: dict) -> None:
         """Update for staged threatconnect data section of profile
 
         This change updates the previous list to a dict with a key that
         can be reference as a variable in other sections of the profile.
 
         Args:
-            profile_data (dict): The current profile data dict.
+            profile_data: The current profile data dict.
 
         Returns:
             dict: The update profile dict.
         """
         if 'stage' not in profile_data:
-            return profile_data
+            return
 
-        stage_tc = profile_data.get('stage').get('threatconnect')
+        stage_tc: Optional[dict] = profile_data.get('stage', {}).get('threatconnect')
 
         # check if profile is using old list type
         if isinstance(stage_tc, list):
@@ -325,14 +337,12 @@ class Migrate:
                 profile_data['stage']['threatconnect'][f'item_{counter}'] = item
                 counter += 1
 
-        return None
-
     @staticmethod
-    def variable_pattern_env_v1(profile_data):
+    def variable_pattern_env_v1(profile_data: dict) -> dict:
         """Update the profile variable to latest pattern
 
         Args:
-            profile_data (dict): The profile data dict.
+            profile_data: The profile data dict.
 
         Returns:
             dict: The updated dict.
@@ -340,8 +350,8 @@ class Migrate:
         profile = json.dumps(profile_data)
 
         for m in re.finditer(r'\"\$(env|envs)\.(\w+)\"', profile):
+            full_match = m.group(0)
             try:
-                full_match = m.group(0)
                 env_type = m.group(1)
                 env_key = m.group(2)
 
@@ -352,11 +362,11 @@ class Migrate:
         return json.loads(profile)
 
     @staticmethod
-    def variable_pattern_env_v2(profile_data):
+    def variable_pattern_env_v2(profile_data: dict) -> dict:
         """Update the profile variable to latest pattern
 
         Args:
-            profile_data (dict): The profile data dict.
+            profile_data: The profile data dict.
 
         Returns:
             dict: The updated dict.
@@ -364,8 +374,8 @@ class Migrate:
         profile = json.dumps(profile_data)
 
         for m in re.finditer(r'\${(env|envs|local|remote)\.(.*?)}', profile):
+            full_match = m.group(0)
             try:
-                full_match = m.group(0)
                 env_type = m.group(1)  # currently env, envs, local, remote
                 env_key = m.group(2)
 
@@ -375,11 +385,11 @@ class Migrate:
                 print(f'{c.Fore.YELLOW}Invalid variable found {full_match}.')
         return json.loads(profile)
 
-    def variable_pattern_tcenv(self, profile_data):
+    def variable_pattern_tcenv(self, profile_data: dict) -> dict:
         """Update the profile variable to latest pattern
 
         Args:
-            profile_data (dict): The profile data dict.
+            profile_data: The profile data dict.
 
         Returns:
             dict: The updated dict.
@@ -396,8 +406,8 @@ class Migrate:
         for data in self.profile.tc_staged_data:
             key = data.get('key')
             for m in re.finditer(r'\${tcenv\.' + str(key) + r'\.(.*?)}', profile):
+                full_match = m.group(0)
                 try:
-                    full_match = m.group(0)
                     jmespath_expression = m.group(1)
 
                     new_variable = f'${{tcenv:{key}:{jmespath_expression}}}'